+++ twiml/voice/verbs/nouns/virtual_agent.go
@@ -1,0 +1,16 @@
+package nouns
+
+import "encoding/xml"
+
+type VirtualAgentAttributes struct {
+	ConnectorName     string  `xml:"connectorName,attr"`
+	Language          *string `xml:"language,attr,omitempty"`
+	SentimentAnalysis *bool   `xml:"sentimentAnalysis,attr,omitempty"`
+	StatusCallback    *string `xml:"statusCallback,attr,omitempty"`
+}
+
+type VirtualAgent struct {
+	XMLName xml.Name `xml:"VirtualAgent"`
+
+	VirtualAgentAttributes
+}